Return a non-zero exit-code on error
This is a common (miss)-behavior of a lot of CLI utilities that rust
makes very easy to fix - by simply returning a Result type from the main()
function (which we already have). This will default to 1 for all
failures.

This is useful for scripting downloads and submissions, but is also
generally the right behavior of a unix utility on error.

We also remove logging the error as this is the default behavior when
returning result from main anyway.

Tested locally with

  $ ./target/debug/aoc d --year 2021 --day 20  # not out yet when tested
  Loaded session cookie from "~/.adventofcode.session".
  Error: "Puzzle 20 of 2021 is still locked."
  $ echo $?
  1